requirements.txt
APScheduler==3.6.3
asgiref==3.3.1
certifi==2020.12.5
dj-database-url==0.5.0
Django==3.1.7
gunicorn==20.1.0
psycopg2==2.8.6
python-telegram-bot==13.4.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
tornado==6.1
tzlocal==2.1
whitenoise==5.2.0
